refactor out catalog service from controller
